misc/python_rest_api_gateway/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
PyYAML==6.0.1
requests==2.31.0
tornado==6.4
urllib3==2.1.0
